fix(components): guard every generated-artifact formatter with the root-prettier check
Review on #1498: `generate-manifest.ts`'s own `formatJson()` imported prettier
independently and never reached the new root-resolution assertion, so a direct
`manifest:generate` / `--check` could still format with a nested shadow copy.
That was one of six independent prettier importers under `scripts/`, every one
of which writes or checks a committed artifact:

  component-artifact-operations.ts  generate-manifest.ts   render-agents-md.ts
  check-promotion-readiness.ts       tokens/generate.ts     tokens/generate-artifacts.ts

The assertion moves to `scripts/lib/prettier-resolution.ts` and is called at each
formatter entry. It checks once and memoizes, so the per-call cost after the
first is a null test. In `generate-artifacts.ts` the registry-JSON call lives
inside an `await Promise.all([...])`, so its guard sits at the statement boundary
above rather than inside the array.

Also narrows the parsed manifest with a type guard in the test (the assertion
from `any` that CI's lint rejected) and fixes the `unknown`-typed comparison it
left behind.

Verified: lint 0 errors, typecheck clean, 317/317 script tests, and every
generator's check passes with the guard live -- components:check, tokens:check,
manifest --check, agents-md --check.
